+++ src/routes.js
@@ -2,7 +2,7 @@ import Dashboard from "./Components/DashboardComponent/Dashboard"
 import Leads from "./Components/LeadsComponent/LeadsComponent"
 import ChatComponent from "./Components/ChatComponent/ChatComponent"
 import LoginForm from "./Components/LoginComponent/LoginForm"
-import UsersComponent from "./Components/AdminPanelComponent/UsersComponent"
+import UsersComponent from "./Components/UsersComponent/UsersComponent"
 import { NoAccess } from "./Components/NoAccess"
 import TaskPage from "./Components/Task/Page/TaskComponent"
 import Schedules from "./Components/Schedules/Schedules"
